Cloudinary Test1

+++ app.js
@@ -16,12 +16,11 @@ const app = express();
 
 app.use(bodyParser.json());
 
-app.use('/uploads/images',express.static(path.join('uploads','images')));
-
 app.use(cors({
   origin: '*',
   methods: ['GET', 'POST', 'DELETE', 'PATCH', 'OPTIONS', 'PUT'],
 }));
+
 app.use('/api/places', placesRoutes); 
 app.use('/api/users', usersRoutes);
 
@@ -31,18 +30,20 @@ app.use((req, res, next) => {
 });
 
 app.use((error, req, res, next) => {
-  if (req.file) {
-    fs.unlink(req.file.path, (err) =>{
+  // Only attempt to unlink local files (not needed for Cloudinary)
+  if (req.file && req.file.path && req.file.path.startsWith('uploads')) {
+    fs.unlink(req.file.path, (err) => {
       console.log(err);
     });
   }
+
   if (res.headerSent) {
     return next(error);
   }
-  res.status(error.code || 500)
-  res.json({message: error.message || 'An unknown error occurred!'});
-});
 
+  res.status(error.code || 500);
+  res.json({ message: error.message || 'An unknown error occurred!' });
+});
 
 mongoose
   .connect(`mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASSWORD}@ac-qvstyix.7jhaea5.mongodb.net/${process.env.DB_NAME}?retryWrites=true&w=majority&appName=Lenscape-users`)
